feat(physics): 💻 split body and quadtree modules
