+++ src/__tests__/app/engineWallFrictionDecay.test.ts
@@ -1,5 +1,6 @@
 /** @jest-environment jsdom */
-import { Engine, Body } from '../../client/physics';
+import { Engine } from '../../client/physics/engine';
+import { Body } from '../../client/physics/body';
 
 describe('Engine wall friction decay', () => {
   it('does not accumulate angular velocity', () => {
